Create Block that can calculate hash
A blockchain comprises of several blocks that are joined to each other.
The chaining of blocks takes place such that if one block is modified,
the rest of the chain become invalid.

Calculate hash will use the attribute values of the object and input it
into a hash algorithm that returns a string representing the contents of
the block and the previous block.

This mean if the previous block is modified, the hash algorithm would
return a different string and invalidate the chain.